+++ mailscanner/functions.php
@@ -21,7 +21,7 @@
 */
 
 // Background colours
-$bg_colors = array("#BDBDBD", "#D8D8D8");
+$bg_colors = array("#EBEBEB", "#D8D8D8");
 
 // Set error level (some distro's have php.ini set to E_ALL)
 if (version_compare(phpversion(), '5.3.0', '<')) {
@@ -602,8 +602,7 @@ function html_start($title, $refresh = 0, $cacheable = true, $report = false)
         echo ' <tr><td>Viruses:</td><td align="right">' . number_format(
                 $row->viruses
             ) . '</td><td align="right">' . $row->viruspercent . '%</tr>' . "\n";
-        echo ' <tr><td>Top Virus:</td><td colspan="2" align="right" style="white-space:nowrap">' . return_todays_top_virus(
-            ) . '</td></tr>' . "\n";
+        echo ' <tr><td>Top Virus:</td><td colspan="2" align="right" style="white-space:nowrap">' . return_todays_top_virus() . '</td></tr>' . "\n";
         echo ' <tr><td>Blocked files:</td><td align="right">' . number_format(
                 $row->blockedfiles
             ) . '</td><td align="right">' . $row->blockedfilespercent . '%</td></tr>' . "\n";
@@ -635,10 +634,14 @@ function html_start($title, $refresh = 0, $cacheable = true, $report = false)
         }
         $nav['reports.php'] = "Reports";
         $nav['other.php'] = "Tools/Links";
-        if ($_SESSION['user_type'] === 'A') {
-            $nav['sf_version.php'] = "Software Versions";
+
+        if (SHOW_SFVERSION == true) {
+            if ($_SESSION['user_type'] === 'A') {
+                $nav['sf_version.php'] = "Software Versions";
+            }
         }
-        if (SHOW_DOC == 'true') {
+
+        if (SHOW_DOC == true) {
             $nav['docs.php'] = "Documentation";
         }
         $nav['logout.php'] = "Logout";
@@ -741,7 +744,6 @@ function row_highandclick()
   }';
 }
 
-
 function html_end($footer = "")
 {
     echo '</td>' . "\n";
@@ -787,8 +789,7 @@ function dbquery($sql)
         //dbtable("SHOW STATUS");
         echo "\n-->\n\n";
     }
-    $result = mysql_query($sql) or die("<B>Error executing query: </B><BR><BR>" . mysql_errno() . ": " . mysql_error(
-        ) . "<BR><BR><B>SQL:</B><BR><PRE>$sql</PRE>");
+    $result = mysql_query($sql) or die("<B>Error executing query: </B><BR><BR>" . mysql_errno() . ": " . mysql_error() . "<BR><BR><B>SQL:</B><BR><PRE>$sql</PRE>");
     return $result;
 }
 
@@ -802,7 +803,6 @@ function quote_smart($value)
     return $value;
 }
 
-
 function safe_value($value)
 {
     dbconn();
@@ -812,7 +812,6 @@ function safe_value($value)
     $value = mysql_real_escape_string($value);
     return $value;
 }
-
 
 function sa_autolearn($spamreport)
 {
@@ -1362,7 +1361,6 @@ function subtract_get_vars($preserve)
     }
 }
 
-
 function subtract_multi_get_vars($preserve)
 {
     if (is_array($_GET)) {
@@ -1388,8 +1386,7 @@ function db_colorised_table($sql, $table_heading = false, $pager = false, $order
 
     // Ordering
     $orderby = null;
-    if (isset($_GET['orderby']))
-    {
+    if (isset($_GET['orderby'])) {
         $orderby = $_GET['orderby'];
         switch (strtoupper($_GET['orderdir'])) {
             case 'A':
@@ -1443,8 +1440,7 @@ function db_colorised_table($sql, $table_heading = false, $pager = false, $order
 
         echo '<table cellspacing="1" class="mail" >
     <tr>
-   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages(
-            ) . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
+   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages() . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
   </tr>
   <tr>
   <td align="center">' . "\n";
@@ -1973,8 +1969,7 @@ function db_colorised_table($sql, $table_heading = false, $pager = false, $order
 
             echo '<table cellspacing="1" class="mail" >
     <tr>
-   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages(
-                ) . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
+   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages() . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
   </tr>
   <tr>
   <td align="center">' . "\n";
@@ -2036,8 +2031,7 @@ function dbtable($sql, $title = false, $pager = false, $operations = false)
 
         echo '<table cellspacing="1" class="mail" >
     <tr>
-   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages(
-            ) . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
+   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages() . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
   </tr>
   <tr>
   <td align="center">' . "\n";
@@ -2128,8 +2122,7 @@ function dbtable($sql, $title = false, $pager = false, $operations = false)
 
         echo '<table cellspacing="1" class="mail" >
     <tr>
-   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages(
-            ) . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
+   <th colspan="5">Displaying page ' . $pager->getCurrentPageID() . ' of ' . $pager->numPages() . ' - Records ' . $from . ' to ' . $to . ' of ' . $pager->numItems() . '</th>
   </tr>
   <tr>
   <td align="center">' . "\n";
@@ -2248,7 +2241,6 @@ function count_files_in_dir($dir)
     return count($array);
 }
 
-
 function get_mail_relays($message_headers)
 {
     $headers = explode("\\n", $message_headers);
@@ -2280,7 +2272,7 @@ function address_filter_sql($addresses, $type)
             break;
         case 'U': // User - show only specific addresses
             foreach ($addresses as $address) {
-                if ((defined('FILTER_TO_ONLY') & FILTER_TO_ONLY)) {
+                if ((defined('FILTER_TO_ONLY') && FILTER_TO_ONLY)) {
                     $sqladdr[] = "to_address like '%$address%'";
                 } else {
                     $sqladdr[] = "to_address like '%$address%' OR from_address = '$address'";
@@ -2292,13 +2284,13 @@ function address_filter_sql($addresses, $type)
         case 'D': // Domain administrator
             foreach ($addresses as $address) {
                 if (strpos($address, '@')) {
-                    if ((defined('FILTER_TO_ONLY') & FILTER_TO_ONLY)) {
+                    if ((defined('FILTER_TO_ONLY') && FILTER_TO_ONLY)) {
                         $sqladdr[] = "to_address like '%$address%'";
                     } else {
                         $sqladdr[] = "to_address like '%$address%' OR from_address = '$address'";
                     }
                 } else {
-                    if ((defined('FILTER_TO_ONLY') & FILTER_TO_ONLY)) {
+                    if ((defined('FILTER_TO_ONLY') && FILTER_TO_ONLY)) {
                         $sqladdr[] = "to_domain='$address'";
                     } else {
                         $sqladdr[] = "to_domain='$address' OR from_domain='$address'";
@@ -2567,10 +2559,6 @@ function quarantine_list($input = "/")
                     $item[] = $f;
                 }
             }
-            if (count($item) > 0) {
-                // Sort in reverse chronological order
-                arsort($item);
-            }
             @closedir($d);
             break;
         default:
@@ -2584,13 +2572,15 @@ function quarantine_list($input = "/")
                             $item[] = "'$f'";
                         }
                     }
-                    if (count($item) > 0) {
-                        asort($item);
-                    }
                     closedir($d);
                 }
             }
     }
+    if (count($item) > 0) {
+        // Sort in reverse chronological order
+        arsort($item);
+    }
+
     return $item;
 }
 
@@ -3103,7 +3093,7 @@ function return_quarantine_dates()
 
 function return_virus_link($virus)
 {
-    if ((defined('VIRUS_INFO') & VIRUS_INFO !== false)) {
+    if ((defined('VIRUS_INFO') && VIRUS_INFO !== false)) {
         $link = sprintf(VIRUS_INFO, $virus);
         return sprintf("<a href=\"%s\">%s</a>", $link, $virus);
     } else {
@@ -3244,4 +3234,3 @@ function funcs_phpversion()
 {
     return (phpversion());
 }
-